backend/requirements.txt
fastapi
uvicorn
pydantic
google-cloud-bigquery
